refactor(PhoneNumberEditor): Move phone verification form to this component.
